requirements.txt
# Core LangChain ecosystem
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0

# LangGraph
langgraph>=0.2.0
langgraph-checkpoint>=2.0.0

# LLM Providers
langchain-google-genai>=2.0.0
langchain-groq>=0.1.0
langchain-huggingface>=0.1.0
langchain-openai>=0.2.0

# Google Gemini API
google-generativeai>=0.8.0

# Hugging Face
huggingface-hub>=0.25.0
transformers>=4.45.0
sentence-transformers>=3.0.0
tokenizers>=0.20.0

# Vector Stores
chromadb>=0.5.0
faiss-cpu>=1.8.0

# Document Processing
pypdf>=4.0.0
unstructured>=0.15.0

# Core dependencies
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.0

# HTTP clients
httpx>=0.27.0
aiohttp>=3.10.0

# Data processing
numpy>=1.26.0
pandas>=2.2.0

# PyTorch (for embeddings/transformers)
torch>=2.4.0

# Utilities
tqdm>=4.66.0
PyYAML>=6.0.0
requests>=2.32.0

# Logging
rich>=13.7.0